# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: BizSawa Backend API
  version: 0.1.0
  description: >
    OpenAPI contract for the BizSawa Go backend.
    It is intended for product learning, API mocking, and test fixture generation.

    Conventions:
    - `Authorization: Bearer <token>` is required for protected routes.
    - Business-scoped routes accept tenant context from the `businessID` path param, JWT claims, or `X-Business-ID`.
    - Mutating requests may accept `X-Idempotency-Key`; it is required for payment initiation.
    - List endpoints use `limit` and `offset` query parameters.
    - Date range filters use RFC3339 timestamps.
    - Money and decimal values are serialized as JSON strings by `shopspring/decimal`.
servers:
  - url: http://localhost:5504
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/profile:
    get:
      tags: [Users]
      summary: Get the current user's profile
      responses:
        "200":
          description: Profile record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      tags: [Users]
      summary: Update the current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    post: 
      tags: [Users]
      summary: Create user's profile under the businessId context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileRequest"
      responses:
        "200":
          description: Created profile record
          content:
            application/json:
              schema:
                $ref: "$/components/schemas/UserProfile"
  /api/v1/businesses:
    get:
      tags: [Business]
      summary: List businesses owned by the current user
      responses:
        "200":
          description: Business list
          content:
            application/json:
              schema:
                type: object
                required: [businesses]
                properties:
                  businesses:
                    type: array
                    items:
                      $ref: "#/components/schemas/Business"
    post:
      tags: [Business]
      summary: Create a new business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBusinessRequest"
      responses:
        "201":
          description: Created business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
...
  /api/v1/businesses/{businessID}/members:
    parameters:
      - $ref: "#/components/parameters/BusinessIDPath"
    get:
      tags: [Users]
      summary: List business members
      responses:
        "200":
          description: Member list
          content:
            application/json:
              schema:
                type: object
                required: [members]
                properties:
                  members:
                    type: array
                    items:
                      $ref: "#/components/schemas/BusinessMember"
    post:
      tags: [Users]
      summary: Invite a member to the business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberRequest"
      responses:
        "201":
          description: Created member invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessMember"
...
  /api/v1/products:
    get:
      tags: [Products]
      summary: List products
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: object
                required: [products]
                properties:
                  products:
                    type: array
                    items:
                      $ref: "#/components/schemas/Product"
    post:
      tags: [Products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductRequest"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
...
  /api/v1/customers:
    get:
      tags: [Customers]
      summary: List customers
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                type: object
                required: [customers]
                properties:
                  customers:
                    type: array
                    items:
                      $ref: "#/components/schemas/Customer"
    post:
      tags: [Customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRequest"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
...
  /api/v1/inventory:
    get:
      tags: [Inventory]
      summary: List inventory items
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                type: object
                required: [inventory]
                properties:
                  inventory:
                    type: array
                    items:
                      $ref: "#/components/schemas/InventoryItem"
...
  /api/v1/orders:
    get:
      tags: [Orders]
      summary: List orders
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: object
                required: [orders]
                properties:
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
    post:
      tags: [Orders]
      summary: Create an order
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyOptional"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /api/v1/sales:
    get:
      tags: [Sales]
      summary: List sales
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Sales list
          content:
            application/json:
              schema:
                type: object
                required: [sales]
                properties:
                  sales:
                    type: array
                    items:
                      $ref: "#/components/schemas/Sale"
    post:
      tags: [Sales]
      summary: Create a sale
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyOptional"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleRequest"
      responses:
        "201":
          description: Created sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
...
  /api/v1/expenses:
    get:
      tags: [Expenses]
      summary: List expenses
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Expense list
          content:
            application/json:
              schema:
                type: object
                required: [expenses]
                properties:
                  expenses:
                    type: array
                    items:
                      $ref: "#/components/schemas/Expense"
    post:
      tags: [Expenses]
      summary: Create an expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseRequest"
      responses:
        "201":
          description: Created expense
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
...
  /api/v1/invoices:
    get:
      tags: [Invoices]
      summary: List invoices
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                type: object
                required: [invoices]
                properties:
                  invoices:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invoice"
    post:
      tags: [Invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
...
  /api/v1/payments:
    get:
      tags: [Payments]
      summary: List payment commands
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Payment command list
          content:
            application/json:
              schema:
                type: object
                required: [payments]
                properties:
                  payments:
                    type: array
                    items:
                      $ref: "#/components/schemas/PaymentCommand"
    post:
      tags: [Payments]
      summary: Initiate a payment command
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyRequired"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiateRequest"
      responses:
        "202":
          description: Created payment command
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentCommand"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BusinessIDPath:
      name: businessID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
...
    IdempotencyKeyOptional:
      name: X-Idempotency-Key
      in: header
      required: false
      schema:
        type: string
    IdempotencyKeyRequired:
      name: X-Idempotency-Key
      in: header
      required: true
      schema:
        type: string
...
  schemas:
...
    UserProfile:
      type: object
      required: [id, tenantId, createdAt, updatedAt, userId, timezone, language]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
        timezone:
          type: string
        language:
          type: string
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
        timezone:
          type: string
        language:
          type: string
    CreateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
        timezone:
          type: string
        language:
          type: string
    Business:
      type: object
      required: [id, tenantId, ownerId, name, slug, currency, timezone, createdAt, updatedAt, mpesaShortcodeConfigured]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        ownerId:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        currency:
          type: string
        timezone:
          type: string
        taxPin:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        mpesaPaymentType:
          type: string
        mpesaShortcodeConfigured:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateBusinessRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        slug:
          type: string
        currency:
          type: string
        timezone:
          type: string
        taxPin:
          type: string
        mpesaPaymentType:
          type: string
          enum: [paybill, pochi_biashara, buy_goods]
        mpesaShortcode:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
...
    BusinessMember:
      type: object
      required: [id, tenantId, createdAt, updatedAt, businessId, userId, role, isActive, invitedAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        businessId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum: [OWNER, MANAGER, CASHIER, VIEWER]
        isActive:
          type: boolean
        invitedBy:
          type: string
          format: uuid
        invitedAt:
          type: string
          format: date-time
        joinedAt:
          type: string
          format: date-time
          nullable: true
    InviteMemberRequest:
      type: object
      required: [userId, role]
      properties:
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum: [MANAGER, CASHIER, VIEWER]
...
    Product:
      type: object
      required: [id, tenantId, createdAt, updatedAt, businessId, name, price, cost, isActive]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        businessId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        sku:
          type: string
        category:
          type: string
        barcode:
          type: string
        imageUrl:
          type: string
        taxRuleId:
          type: string
          format: uuid
          nullable: true
        price:
          type: string
        cost:
          type: string
        isActive:
          type: boolean
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
    ProductVariant:
      type: object
      required: [id, tenantId, createdAt, updatedAt, businessId, productId, name, price, cost, isActive]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        businessId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        name:
          type: string
        sku:
          type: string
        barcode:
          type: string
        price:
          type: string
        cost:
          type: string
        isActive:
          type: boolean
    VariantRequest:
      type: object
      properties:
        name:
          type: string
        sku:
          type: string
        barcode:
          type: string
        price:
          type: string
        cost:
          type: string
        isActive:
          type: boolean
    ProductRequest:
      type: object
      required: [name, price, cost]
      properties:
        name:
          type: string
        description:
          type: string
        sku:
          type: string
        category:
          type: string
        barcode:
          type: string
        imageUrl:
          type: string
        taxRuleId:
          type: string
          format: uuid
          nullable: true
        price:
          type: string
        cost:
          type: string
        isActive:
          type: boolean
        variants:
          type: array
          items:
            $ref: "#/components/schemas/VariantRequest"
    Customer:
      type: object
      required: [id, tenantId, createdAt, updatedAt, businessId, name, loyaltyPoints, totalSpend]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        businessId:
          type: string
          format: uuid
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
        loyaltyPoints:
          type: integer
        totalSpend:
          type: string
        lastPurchaseAt:
          type: string
          format: date-time
          nullable: true
    CustomerRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
        loyaltyPoints:
          type: integer
        totalSpend:
          type: string
...
    InventoryItem:
      type: object
      required: [id, tenantId, createdAt, updatedAt, businessId, productId, quantity, lowStockThreshold]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        businessId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        quantity:
          type: string
        lowStockThreshold:
          type: string
...
    Order:
      type: object
      required: [id, tenantId, createdAt, updatedAt, businessId, status, subtotal, taxAmount, total, paymentMethod]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        businessId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [draft, confirmed, fulfilled, cancelled, refunded]
        subtotal:
          type: string
        taxAmount:
          type: string
        total:
          type: string
        paymentMethod:
          type: string
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        fulfilledAt:
          type: string
          format: date-time
          nullable: true
        lines:
          type: array
          items:
            $ref: "#/components/schemas/OrderLine"
    OrderLine:
      type: object
      required: [id, tenantId, createdAt, updatedAt, businessId, orderId, productId, quantity, unitPrice, lineTotal]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        businessId:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        quantity:
          type: string
        unitPrice:
          type: string
        lineTotal:
          type: string
    OrderLineRequest:
      type: object
      required: [productId, quantity, unitPrice]
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: string
        unitPrice:
          type: string
    CreateOrderRequest:
      type: object
      required: [lines]
      properties:
        customerId:
          type: string
          format: uuid
          nullable: true
        paymentMethod:
          type: string
        lines:
          type: array
          items:
            $ref: "#/components/schemas/OrderLineRequest"
    Sale:
      type: object
      required: [id, tenantId, createdAt, updatedAt, businessId, receiptNumber, staffId, paymentMethod, subtotal, taxAmount, total, status, soldAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        businessId:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
          nullable: true
        customerId:
          type: string
          format: uuid
          nullable: true
        receiptNumber:
          type: string
        staffId:
          type: string
          format: uuid
        paymentMethod:
          type: string
        subtotal:
          type: string
        taxAmount:
          type: string
        total:
          type: string
        status:
          type: string
        soldAt:
          type: string
          format: date-time
        lines:
          type: array
          items:
            $ref: "#/components/schemas/SaleLine"
...
    SaleLineRequest:
      type: object
      required: [productId, quantity, unitPrice]
      properties:
        productId:
          type: string
          format: uuid
        quantity:
          type: string
        unitPrice:
          type: string
    CreateSaleRequest:
      type: object
      required: [lines]
      properties:
        orderId:
          type: string
          format: uuid
          nullable: true
        customerId:
          type: string
          format: uuid
          nullable: true
        paymentMethod:
          type: string
        lines:
          type: array
          items:
            $ref: "#/components/schemas/SaleLineRequest"
...
    Expense:
      type: object
      required: [id, tenantId, createdAt, updatedAt, businessId, category, amount, taxAmount, isRecurring, spentAt]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        businessId:
          type: string
          format: uuid
        category:
          type: string
        description:
          type: string
        vendor:
          type: string
        amount:
          type: string
        taxAmount:
          type: string
        isRecurring:
          type: boolean
        recurringInterval:
          type: string
        spentAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
    ExpenseRequest:
      type: object
      required: [category, amount]
      properties:
        category:
          type: string
        description:
          type: string
        vendor:
          type: string
        amount:
          type: string
        taxAmount:
          type: string
        isRecurring:
          type: boolean
        recurringInterval:
          type: string
        spentAt:
          type: string
          format: date-time
          nullable: true
...
    Invoice:
      type: object
      required: [id, tenantId, createdAt, updatedAt, businessId, invoiceNumber, status, subtotal, taxAmount, total, amountPaid, amountDue, currency]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        businessId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
          nullable: true
        invoiceNumber:
          type: string
        status:
          type: string
          enum: [draft, sent, viewed, partial, paid, overdue, cancelled]
        subtotal:
          type: string
        taxAmount:
          type: string
        total:
          type: string
        amountPaid:
          type: string
        amountDue:
          type: string
        currency:
          type: string
        notes:
          type: string
        dueAt:
          type: string
          format: date-time
          nullable: true
        sentAt:
          type: string
          format: date-time
          nullable: true
        viewedAt:
          type: string
          format: date-time
          nullable: true
        paidAt:
          type: string
          format: date-time
          nullable: true
        lines:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLine"
...
    InvoiceLineRequest:
      type: object
      required: [description, quantity, unitPrice]
      properties:
        productId:
          type: string
          format: uuid
          nullable: true
        description:
          type: string
        quantity:
          type: string
        unitPrice:
          type: string
    CreateInvoiceRequest:
      type: object
      required: [lines]
      properties:
        customerId:
          type: string
          format: uuid
          nullable: true
        currency:
          type: string
        notes:
          type: string
        dueAt:
          type: string
          format: date-time
          nullable: true
        lines:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineRequest"
...
    PaymentCommand:
      type: object
      required: [id, tenantId, createdAt, updatedAt, businessId, type, status, amount, currency, provider]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        businessId:
          type: string
          format: uuid
        invoiceId:
          type: string
          format: uuid
          nullable: true
        type:
          type: string
          enum: [stk_push, b2c, c2b]
        status:
          type: string
          enum: [pending, processing, succeeded, failed]
        amount:
          type: string
        currency:
          type: string
        phone:
          type: string
        accountReference:
          type: string
        provider:
          type: string
        providerRequestId:
          type: string
        providerReceipt:
          type: string
        failureCode:
          type: string
        failureMessage:
          type: string
        payload:
          type: object
          additionalProperties: true
        resultPayload:
          type: object
          additionalProperties: true
        processedAt:
          type: string
          format: date-time
          nullable: true
    InitiateRequest:
      type: object
      required: [amount]
      properties:
        type:
          type: string
          enum: [stk_push, b2c, c2b]
        invoiceId:
          type: string
          format: uuid
          nullable: true
        amount:
          type: string
        currency:
          type: string
        phone:
          type: string
        accountReference:
          type: string
        payload:
          type: object
          additionalProperties: true